split useCustomClaims into its own file
