+++ src/components/index.ts
@@ -1,5 +1,6 @@
 import { Impersonation } from './impersonation.js';
 import { AuthKitProvider, useAuth } from './authkit-provider.js';
-import { useAccessToken, useCustomClaims } from './useAccessToken.js';
+import { useAccessToken } from './useAccessToken.js';
+import { useCustomClaims } from './useCustomClaims.js';
 
 export { Impersonation, AuthKitProvider, useAuth, useAccessToken, useCustomClaims };
